pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["Dirlan Fernando <dirlanfr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"
pytest-django = "^4.9.0"
faker = "^30.8.2"
factory-boy = "^3.3.1"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
